feat(Create Unical Cronofy connector): Refactored file structure. [sc-796]
